refactored directory layout into a src/ and lib/ layout; added knife library as a dependency via git submodules; refactored Bar and Legend to extend knife's Base class
