+++ src/mbm/screens/playing/bullet.lua
@@ -1,4 +1,4 @@
-local check = require "check-self"
+local check = require "mbm.shared.check-self"
 
 ---@class Bullet # The Bullet class.
 local Bullet = {
