+++ src/mbm/screens/playing/hit-effect.lua
@@ -1,4 +1,4 @@
-local check = require "check-self"
+local check = require "mbm.shared.check-self"
 
 ---@class HitEffect
 local HitEffect = {
