+++ src/mbm/screens/playing/hit-effects.lua
@@ -1,4 +1,4 @@
-local check = require "check-self"
+local check = require "mbm.shared.check-self"
 
 ---@class HitEffects # The collection of hit effects.
 local HitEffects = {
