+++ src/mbm/shared/check-self.lua
