+++ src/mbm/shared/clip.lua
