+++ src/mbm/shared/ground.lua
@@ -1,4 +1,4 @@
-local check = require "check-self"
+local check = require "mbm.shared.check-self"
 
 ---@class Ground # The ground.
 local Ground = {
