+++ src/mbm/shared/moon.lua
@@ -1,4 +1,4 @@
-local check = require "check-self"
+local check = require "mbm.shared.check-self"
 
 ---@class Moon # The Moon.
 local Moon = {
